fix input capture

+++ include/drivers/counter/sam_tc.h
@@ -10,7 +10,7 @@
 #include <device.h>
 #include <drivers/counter.h>
 
-int sam_counter_read_capture(const struct device *dev, uint16_t *period);
+int sam_counter_read_capture(const struct device *dev, uint32_t *period);
 
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_COUNTER_SAM_TC_H_ */
